requirements.txt
flask==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
pandas==2.2.2
requests==2.32.3
sqlalchemy==2.0.30
